after 15k years...

+++ include/Kydo/Window.h
@@ -22,7 +22,7 @@ namespace Kydo
 		PCWSTR className, title;
 		WNDCLASSEXW clazz;
 
-		SIZE_T nPixels;
+		LONG nPixels;
 		LONG x, y, width, height;
 		COLORREF *pixels;
 
@@ -46,7 +46,7 @@ namespace Kydo
 		LONG GetHeight();
 
 		COLORREF *GetPixels();
-		SIZE_T GetPixelCount();
+		LONG GetPixelCount();
 
 		KYDO_PROPERTY(get = GetX) LONG X;
 		KYDO_PROPERTY(get = GetY) LONG Y;
@@ -54,7 +54,7 @@ namespace Kydo
 		KYDO_PROPERTY(get = GetHeight) LONG Height;
 
 		KYDO_PROPERTY(get = GetPixels) COLORREF *Pixels;
-		KYDO_PROPERTY(get = GetPixelCount) SIZE_T PixelCount;
+		KYDO_PROPERTY(get = GetPixelCount) LONG PixelCount;
 
 		static LRESULT CALLBACK KydoWinProc(__in HWND wnd, __in UINT msg, __in WPARAM wParam, __in LPARAM lParam);
 	};
